refactor(ui): add support for fullscreen profile layout
