+++ packages/ui/src/pages/Profile/index.tsx
@@ -1,0 +1,5 @@
+const Profile = () => {
+  return <>Profile works!</>;
+};
+
+export default Profile;
